requirements.txt
# Core dependencies for Legal Contract Analyzer
flask==2.3.3
pdfplumber==0.10.3
python-docx==1.1.0
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.0
spacy==3.7.2
nltk==3.8.1
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0

# Optional: For advanced NLP and LLM integration
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2

# Development dependencies
pytest==7.4.2
black==23.9.1
flake8==6.1.0 